Create PendingWithdrawalView component
We created separated PendingWithdrawalView component, and in PendingWithdrawals
component we only keep submit functions. This will make PendingWithdrawal
component more reusable (there will be no need to create the submit functions
each time we want to use it).
